+++ src/crc/crc8/crc8.rs
@@ -35,7 +35,7 @@ mod consts {
 }
 
 /// The CRC16 hasher
-#[derive(Copy, Clone)]
+#[derive(Copy, Clone, Debug)]
 pub struct Crc8 {
     state: u8,
 }