+++ src/macros/utils.rs
@@ -9,7 +9,7 @@ macro_rules! implement_digest {
     ($Hasher:ident, $BlockSize:ident, $OutputSize:ident) => {
         #[cfg(feature = "generic")]
         impl digest::BlockInput for $Hasher {
-            type BlockSize = generic_array::typenum::$BlockSize;
+            type BlockSize = digest::generic_array::typenum::$BlockSize;
         }
 
         #[cfg(feature = "generic")]
@@ -22,11 +22,11 @@ macro_rules! implement_digest {
 
         #[cfg(feature = "generic")]
         impl digest::FixedOutput for $Hasher {
-            type OutputSize = generic_array::typenum::$OutputSize;
+            type OutputSize = digest::generic_array::typenum::$OutputSize;
             #[inline]
             fn fixed_result(self) -> generic_array::GenericArray<u8, Self::OutputSize> {
                 use generic_array::typenum::Unsigned;
-                let mut array = generic_array::GenericArray::default();
+                let mut array = digest::generic_array::GenericArray::default();
                 let mut out = self.finish();
                 let size = Self::OutputSize::to_usize();
                 for i in 0..size {